polish(lisong): polish code
