+++ dizoo/beergame/config/beergame_onppo_config.py
@@ -5,13 +5,13 @@ beergame_ppo_config = dict(
     env=dict(
         collector_env_num=8,
         evaluator_env_num=8,
-        # env_id='beergame-v2',
         n_evaluator_episode=8,
         stop_value=200,
         role=0,  # 0-3 : retailer, warehouse, distributor, manufacturer
         agent_type='bs',
-        replay_path=None
-        # type of co-player, 'bs'- base stock, 'Strm'- use Sterman formula to model typical human behavior.
+        # type of co-player, 'bs'- base stock, 'Strm'- use Sterman formula to model typical human behavior
+        demandDistribution=0
+        # distribution of demand, default=0, '0=uniform, 1=normal distribution, 2=the sequence of 4,4,4,4,8,..., 3= basket data, 4= forecast data'
     ),
     policy=dict(
         cuda=True,
@@ -29,9 +29,7 @@ beergame_ppo_config = dict(
             epoch_per_collect=10,
             batch_size=320,
             learning_rate=3e-4,
-            value_weight=0.5,
             entropy_weight=0.001,
-            clip_ratio=0.2,
             adv_norm=True,
             value_norm=True,
             # for onppo, when we recompute adv, we need the key done in data to split traj, so we must
@@ -39,12 +37,9 @@ beergame_ppo_config = dict(
             # but when we add key traj_flag in data as the backup for key done, we could choose to use ignore_done=True
             # for halfcheetah, the length=1000
             ignore_done=True,
-            grad_clip_type='clip_norm',
-            grad_clip_value=0.5,
         ),
         collect=dict(
             n_episode=8,
-            unroll_len=1,
             discount_factor=0.99,
             gae_lambda=0.95,
             collector=dict(
@@ -64,16 +59,12 @@ beergame_ppo_create_config = dict(
     ),
     env_manager=dict(type='base'),
     policy=dict(type='ppo'),
-    collector=dict(
-        type='episode',
-        get_train_sample=True,
-        reward_shaping=True,  # whether use total reward to reshape reward
-    ),
+    collector=dict(type='episode', ),
 )
 beergame_ppo_create_config = EasyDict(beergame_ppo_create_config)
 create_config = beergame_ppo_create_config
 
 if __name__ == "__main__":
-    # or you can enter `ding -m serial -c beergame_offppo_config.py -s 0`
+    # or you can enter `ding -m serial_onpolicy -c beergame_onppo_config.py -s 0`
     from ding.entry import serial_pipeline_onpolicy
     serial_pipeline_onpolicy([main_config, create_config], seed=0)
